Create Java-Substring
Solution to substrings in Java. substring() is a method is java that helps us fetch a substring that is included within a string. If we specify 2 indices within the brackets, the method will fetch substring between the specified indices. However, if we specify only a single index then the method shall fetch substring from the specified index till the end of the string. 

NOTE: The first index is inclusive and the second index is exclusive.