Create Java End-Of-file
In computing, End Of File (commonly abbreviated EOF) is a condition in a computer operating system where no more data can be read from a data source. We use the hasNextLine() method of the scanner class to check whether or not there are any further lines to read and if not, we determine that we have reached the file's end.